requirements-etl.txt
asyncpg==0.30.0
httpx==0.27.2
pandas==2.2.3
python-dotenv==1.0.1
alembic==1.13.3
psycopg2-binary==2.9.10
